+++ scripts/scene_elements/camera.py
@@ -129,3 +129,4 @@ class Camera:
         size.y = abs(size.y)
         offset = self._pos - (size / 2)
         return pygame.Vector2(round(offset[0]), round(offset[1]))
+
