+++ scripts/scene_elements/world/view.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import Optional, TYPE_CHECKING
 
 import pygame
 
@@ -82,7 +82,7 @@ class WorldView:
         """
         Update the camera's position to follow the player's Units
         """
-        target_pos = self._game.memory.get_team_center("player")
+        target_pos = self.get_team_center("player")
         if target_pos:
             self.camera.move_to_position(target_pos)
             self.camera.update(delta_time)
@@ -123,3 +123,18 @@ class WorldView:
                         for p in ([entity.pos] + entity.behaviour.current_path)
                     ]
                     pygame.draw.lines(surface, (255, 0, 0), False, points)
+
+    def get_team_center(self, team) -> Optional[pygame.Vector2]:
+        """
+        Get centre coordinates for the team
+        """
+        count = 0
+        pos_totals = pygame.Vector2()
+        for unit in self._game.memory.get_all_units():
+            if unit.team == team:
+                pos_totals += unit.pos
+                count += 1
+        if count:
+            return pygame.Vector2(pos_totals / count)
+        else:
+            return None
